titan_intra_service_auth/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "titan-intra-service-auth"
version = "6.0.0"
description = "Microserviço de autenticação intra-serviços — Titan Intra Service Auth Engine"
readme = "README.md"
authors = [{ name = "Elias Andrade", email = "Replika AI Solutions" }]
license = { text = "Proprietary" }
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
    "psutil>=5.9.0",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21", "httpx>=0.25"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
